Project Colors moved to YaProjectEditor
